gdb: add some extra debug information to attach_command
While working on another patch I wanted to add some extra debug
information to the attach_command function.  This required me to add a
new function to convert the thread_info::state variable to a string.

The new debug might be useful to others, and the state to string
function might be useful in other locations, so I thought I'd merge
it.